Refactor: Move the custom runtime loader to a separate file
